requirements.txt
# Core framework
Django==5.0.3

# Django REST Framework
djangorestframework==3.15.1

# Gunicorn for production WSGI server
gunicorn==21.2.0

# PostgreSQL adapter
psycopg2-binary==2.9.9

# Optional: environment variable management
python-decouple==3.8

# Optional: CORS headers if you need cross-origin API access
django-cors-headers==4.3.1